+++ Toolbox/mylab/envs/go_explore_env.py
@@ -11,228 +11,6 @@ import time
 from cached_property import cached_property
 from garage.core import Parameterized
 
-
-
-class CellPool():
-    def __init__(self, filename = 'database.dat', flag=db.DB_RDONLY, flag2='r'):
-        # print("Creating new Cell Pool:", self)
-        # self.guide = set()
-
-        # import pdb; pdb.set_trace()
-        # self.pool = [self.init_cell]
-        # self.guide = self.init_cell.observation
-        self.length = 0
-
-        # self.d_pool = {}
-
-        pool_DB = db.DB()
-        # print('Creating Cell Pool with flag:', flag)
-        # print(filename)
-        pool_DB.open(filename, dbname=None, dbtype=db.DB_HASH, flags=flag)
-        # pool_DB = None
-        # self.d_pool = shelve.Shelf(pool_DB, protocol=pickle.HIGHEST_PROTOCOL)
-        self.key_list = []
-        self.max_value = 0
-        self.max_score = 0
-        # self.d_pool = shelve.BsdDbShelf(pool_DB)
-        # self.d_pool = shelve.open('/home/mkoren/Scratch/cellpool-shelf2', flag=flag2)
-        # self.d_pool = shelve.DbfilenameShelf('/home/mkoren/Scratch/cellpool-shelf2', flag=flag2)
-
-    def create(self, d_pool):
-
-        self.init_cell = Cell()
-        self.init_cell.observation = np.zeros((1,128))
-        self.init_cell.trajectory = None
-        self.init_cell.score = -np.inf
-        self.init_cell.state = None
-        self.init_cell.times_chosen = 0
-        self.init_cell.times_visited = 1
-        # self.d_pool = shelve.open('cellpool-shelf', flag=flag)
-
-        d_pool[str(hash(self.init_cell))] = self.init_cell
-        self.key_list.append(str(hash(self.init_cell)))
-        self.length = 1
-        self.max_value = self.init_cell.fitness
-        # import pdb; pdb.set_trace()
-
-    # def append(self, cell):
-    #     # pdb.set_trace()
-    #     # if observation not in self.guide:
-    #     #     self.guide.add(observation)
-    #     #     cell = Cell()
-    #     #     cell.observation = observation
-    #     #     self.pool.append(cell)
-    #     #     self.length += 1
-    #     if cell in self.d_pool:
-    #         self.d_pool[cell].seen += 1
-    #     else:
-    #         self.d_pool[cell] = cell
-
-
-    # def get_cell(self, index):
-    #     return self.pool[index]
-    #
-    # def get_random_cell(self):
-    #     index = np.random.randint(0, self.length)
-    #     return self.get_cell(index)
-
-    def d_update(self, d_pool, observation, trajectory, score, state, chosen=0):
-        # pdb.set_trace()
-        #This tests to see if the observation is already in the matrix
-        obs_hash = str(hash(observation.tostring()))
-        if not obs_hash in d_pool:
-            # self.guide.add(observation)
-            cell = Cell()
-            cell.observation = observation
-            # self.guide = np.append(self.guide, np.expand_dims(observation, axis=0), axis = 0)
-            cell.trajectory = trajectory
-            cell.score = score
-            cell.trajectory_length = len(trajectory)
-            cell.state = state
-            cell.times_visited = 1
-            cell.times_chosen = chosen
-            cell.times_chosen_since_improved = 0
-            d_pool[obs_hash] = cell
-            self.length += 1
-            self.key_list.append(obs_hash)
-            if cell.fitness > self.max_value:
-                self.max_value = cell.fitness
-            if cell.score > self.max_score:
-                self.max_score = score
-            return True
-        else:
-            cell = d_pool[obs_hash]
-            if score > cell.score:
-                cell.score = score
-                cell.trajectory = trajectory
-                cell.trajectory_length = len(trajectory)
-                cell.state = state
-
-            cell.times_visited += 1
-            cell.times_chosen += chosen
-            d_pool[obs_hash] = cell
-            if cell.fitness > self.max_value:
-                self.max_value = cell.fitness
-            if cell.score > self.max_score:
-                self.max_score = score
-        return False
-
-
-class Cell():
-
-    def __init__(self):
-        # print("Creating new Cell:", self)
-        # Number of times this was chosen and seen
-        self._times_visited=0
-        self._times_chosen = 0
-        self._times_chosen_since_improved = 0
-        self._score = -np.inf
-        self._action_times = 0
-
-        self.trajectory_length = -np.inf
-        self.trajectory = []
-        self.state = None
-        self.observation = None
-
-    def __eq__(self, other):
-        if type(other) != type(self):
-            return False
-        if np.all(self.observation == other.observation):
-            return True
-        else:
-            return False
-
-    def reset_cached_property(self, cached_property):
-        if cached_property in self.__dict__:
-            del self.__dict__[cached_property]
-
-    @property
-    def score(self):
-        return self._score
-
-    @score.setter
-    def score(self, value):
-        self._score = value
-        self.reset_cached_property('score_weight')
-        self.reset_cached_property('fitness')
-
-    @property
-    def times_visited(self):
-        return self._times_visited
-
-    @times_visited.setter
-    def times_visited(self, value):
-        self._times_visited = value
-        self.reset_cached_property('times_visited_subscore')
-        self.reset_cached_property('count_subscores')
-        self.reset_cached_property('fitness')
-
-    @property
-    def times_chosen(self):
-        return self._times_chosen
-
-    @times_chosen.setter
-    def times_chosen(self, value):
-        self._times_chosen = value
-        self.reset_cached_property('times_chosen_subscore')
-        self.reset_cached_property('count_subscores')
-        self.reset_cached_property('fitness')
-
-    @property
-    def times_chosen_since_improved(self):
-        return self._times_chosen_since_improved
-
-    @times_chosen_since_improved.setter
-    def times_chosen_since_improved(self, value):
-        self._times_chosen_since_improved = value
-        self.reset_cached_property('times_chosen_since_improved')
-        self.reset_cached_property('count_subscores')
-        self.reset_cached_property('fitness')
-
-
-    @cached_property
-    def fitness(self):
-        # return max(1, self.score)
-        return self.score_weight*(self.count_subscores + 1)
-
-    @cached_property
-    def count_subscores(self):
-        return (self.times_chosen_subscore +
-                self.times_chosen_since_improved_subscore +
-                self.times_visited_subscore)
-
-    @cached_property
-    def times_chosen_subscore(self):
-        weight = 0.1
-        power = 0.5
-        eps1 = 0.001
-        eps2 = 0.00001
-        return weight * (1 / (self.times_chosen + eps1)) ** power + eps2
-
-    @cached_property
-    def times_chosen_since_improved_subscore(self):
-        weight = 0.0
-        power = 0.5
-        eps1 = 0.001
-        eps2 = 0.00001
-        return weight * (1 / (self.times_chosen_since_improved + eps1)) ** power + eps2
-
-    @cached_property
-    def times_visited_subscore(self):
-        weight = 0.3
-        power = 0.5
-        eps1 = 0.001
-        eps2 = 0.00001
-        return weight * (1 / (self.times_chosen + eps1)) ** power + eps2
-
-    @cached_property
-    def score_weight(self):
-        return 1.0
-        # return min(1e-6, 0.1**max(0.0, (100000-self.score)/10000))
-
-    def __hash__(self):
-        return hash((self.observation.tostring()))
-
 class GoExploreParameter():
     def __init__(self, name, value, **tags):
         self.name = name
@@ -249,23 +27,15 @@ class GoExploreParameter():
 
 
 class GoExploreTfEnv(TfEnv, Parameterized):
-    # cell_pool = CellPool()
-    # pool = []
-    # var = Value('i', 7)
+
     def __init__(self, env=None, env_name=""):
         self.params_set = False
         self.db_filename = 'database.dat'
         self.key_list = []
         self.max_value = 0
-        # self.downsampler = self.default_downsampler
 
         super().__init__(env, env_name)
         Parameterized.__init__(self)
-        # self.cell_pool = cell_pool
-
-        # print("New env, pool: ", GoExploreTfEnv.pool)
-        # print("New env: ", self, " test_var: ", self.test_var)
-        # print("init object: ", GoExploreTfEnv)
 
     def sample(self, population):
         #Proportional sampling: Stochastic Acceptance
@@ -291,18 +61,7 @@ class GoExploreTfEnv(TfEnv, Parameterized):
 
         Calls reset on wrapped env.
         """
-        # print("reset")
-        # print(self.p_db_filename.value)
-        # flag = db.DB_RDONLY
-        # pool_DB = db.DB()
-        # pool_DB.open(self.p_db_filename.value, dbname=None, dbtype=db.DB_HASH, flags=flag)
-        # dd_pool = shelve.Shelf(pool_DB)
-        # cell = dd_pool[random.choice(list(dd_pool.keys()))]
-        # dd_pool.close()
-        # pool_DB.close()
-        # pdb.set_trace()
-        # self.env.env.restore_state(cell.state)
-        # obs = self.env.env._get_obs()
+
         try:
             # pdb.set_trace()
             # start = time.time()
@@ -376,29 +135,8 @@ class GoExploreTfEnv(TfEnv, Parameterized):
             print("Failed to get state from database")
             obs = self.env.env.reset()
 
-        # o = super().reset(**kwargs)
-        # print("In reset")
-        # if self.p_pool is not None: print("Cell Pool Length: ", self.p_pool.get_value().length)
 
-        #TODO: This will be slow, need to implement faster random item dict (memory trade off)
-        #TODO: Also need to sample cells proportional to reward
-        #https://stackoverflow.com/questions/2140787/select-k-random-elements-from-a-list-whose-elements-have-weights
-        # cell = self.p_pool.value.d_pool[random.choice(list(self.p_pool.value.d_pool.keys()))]
-        # if cell.state is not None:
-        #     # pdb.set_trace()
-        #     self.env.env.restore_state(cell.state)
-        #     obs = self.env.env._get_obs()
-        # else:
-        #     obs = self.env.env.reset()
-        # obs = self.env.env._get_obs()
-        # print("Got obs")
-        # cell = self.cell_pool.get_random_cell()
-        # if cell.state is None:
-        #     return super().reset(**kwargs)
-        # self.env.restore_state(cell.state)
-        # print("downsample")
         x = self.downsample(obs)
-        # print("return from reset")
         return x
     #
     def step(self, action):
@@ -408,8 +146,7 @@ class GoExploreTfEnv(TfEnv, Parameterized):
 
         Calls step on wrapped env.
         """
-        # import pdb; pdb.set_trace()
-        # print("step")
+
         try:
             obs, reward, done, env_info = self.env.env.step(action)
             env_info['state'] = self.env.env.clone_state()
@@ -420,29 +157,14 @@ class GoExploreTfEnv(TfEnv, Parameterized):
         except:
             pdb.set_trace()
 
-    # def set_cell_pool(self, cell_pool):
-    #     self.cell_pool = cell_pool
-    #     print(self, "had cell pool set to: ", self.cell_pool)
-
-    # def append_cell(self, cell):
-        # GoExploreTfEnv.pool.append(cell)
-        # print("Appended Cell: ", cell, " -- pool length: ", len(GoExploreTfEnv.pool))
-        # print("append object: ", GoExploreTfEnv)
-        # print("appending env with pool: ", GoExploreTfEnv.var.value)
-        # GoExploreTfEnv.var.value = np.random.randint(0, 100)
-        # print("appending env with mopdified pool: ", GoExploreTfEnv.var.value)
 
     @overrides
     def get_params_internal(self, **tags):
         # this lasagne function also returns all var below the passed layers
         if not self.params_set:
-            # self.p_pool = GoExploreParameter("pool", CellPool())
             self.p_db_filename = GoExploreParameter("db_filename", self.db_filename)
             self.p_key_list = GoExploreParameter("key_list", self.key_list)
             self.p_max_value = GoExploreParameter("max_value", self.max_value)
-            # self.p_downsampler = GoExploreParameter("downsampler", self.downsampler)
-            # self.p_var = GoExploreParameter("var", GoExploreTfEnv.var)
-            # self.p_test_var = GoExploreParameter("test_var", self.test_var)
             self.params_set = True
 
         if tags.pop("db_filename", False) == True:
@@ -454,37 +176,20 @@ class GoExploreTfEnv(TfEnv, Parameterized):
         if tags.pop("max_value", False) == True:
             return [self.p_max_value]
 
-        # if tags.pop("downsampler", False) == True:
-            # return [self.p_downsampler]
 
         return [self.p_db_filename, self.p_key_list, self.p_max_value]#, self.p_downsampler]
 
-        # if tags.pop("pool", False) == True:
-        #     return [self.p_pool]
-        # if tags.pop("test_var", False) == True:
-            # return [self.p_test_var]
-        # if tags.pop("pool", False) == True:
-        #     return [self.p_pool]
-        # return [self.p_pool]
-        # return [self.p_pool,self.p_var, self.p_test_var]
-
     @overrides
     def set_param_values(self, param_values, **tags):
-        # pdb.set_trace()
-        # if tags['pool'] is not None:
-        #     GoExploreTfEnv.pool = tags['pool']
-        #     print("set pool")
 
         debug = tags.pop("debug", False)
-        # param_values = unflatten_tensors(flattened_params,
-        #                                  self.get_param_shapes(**tags))
+
         for param,  value in zip(
             self.get_params(**tags),
             param_values):
             param.set_value(value)
             if debug:
                 print("setting value of %s" % param.name)
-        # super().set_param_values(flattened_params, **tags)
 
     @overrides
     def get_param_values(self, **tags):
@@ -495,28 +200,6 @@ class GoExploreTfEnv(TfEnv, Parameterized):
     def downsample(self, obs):
         return obs
 
-    # def downsample(self, obs):
-    #     return self.downsampler(obs=obs)
-    #
-    # def default_downsampler(self, obs):
-    #     print("DEFAULT DOWNSAMPLE")
-    #     return obs
-from skimage.measure import block_reduce
-
-class Pixel_GoExploreEnv(GoExploreTfEnv):
-    @overrides
-    def downsample(self, obs):
-        # import pdb; pdb.set_trace()
-        obs = np.dot(obs[..., :3], [0.299, 0.587, 0.114])
-        obs = block_reduce(obs, block_size=(20, 20), func=np.mean)
-        obs= obs.astype(np.uint8) // 32
-        return obs.flatten()
-
-class Ram_GoExploreEnv(GoExploreTfEnv):
-    @overrides
-    def downsample(self, obs):
-        # import pdb; pdb.set_trace()
-        return obs // 32
 
 
 # Might be able to speed up by having class of (observation, index) where hash and eq
